Changing the project name
