--- APP/README.md
@@ -1,2 +1,0 @@
-# ProjetoIntegrado
-Repositório contendo códigos desenvolvidos para a disciplina de Projeto Integrado
