--- APP/app/src/main/java/com/software/sadv/MainActivity.java
@@ -1,15 +1,0 @@
-package com.software.sadv;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        int a;
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
