+++ StreetAnalyzer/app/.gitignore
