+++ StreetAnalyzer/app/src/main/java/com/street/analyzer/MainActivity.java
@@ -1,0 +1,14 @@
+package com.street.analyzer;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
