+++ StreetAnalyzer/gradle.properties
