+++ StreetAnalyzer/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Apr 09 10:26:57 BRT 2020
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.6.4-all.zip
